+++ tests/Feature/Auth/AuthenticationTest.php
@@ -21,7 +21,7 @@ class AuthenticationTest extends TestCase
 
         $response = $this->post(route('login'), [
             'email' => $user->email,
-            'password' => 'password',
+            'password' => 'p@ssWORD1234',
         ]);
 
         $this->assertAuthenticated();